requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
paho-mqtt>=1.6.1
django-cors-headers>=4.3.0